refactor: Extract required fields logic to helpers
