+++ packages/cozy-harvest-lib/src/helpers/manifest.js
@@ -3,6 +3,7 @@ import _cloneDeep from 'lodash/cloneDeep'
 import findKey from 'lodash/findKey'
 import _mapValues from 'lodash/mapValues'
 import _pickBy from 'lodash/pickBy'
+import intersection from 'lodash/intersection'
 
 export const ROLE_IDENTIFIER = 'identifier'
 
@@ -213,10 +214,39 @@ const getFieldsValues = (konnector, account) => {
   return initialAndDefaultValues
 }
 
+/**
+ * Get required fields names for a given konnector
+ *
+ * @param {Object} konnector - an io.cozy.konnectors document
+ *
+ * @returns {string[]} An array of all required fields names
+ */
+const getRequiredFields = konnector => {
+  const { fields } = konnector
+  const sanitizedFields = sanitizeFields(fields)
+
+  const requiredFields = Object.entries(sanitizedFields)
+    .filter(([, value]) => value.required)
+    .map(([key]) => key)
+
+  return requiredFields
+}
+
+const hasValuesForRequiredFields = (konnector, values) => {
+  const requiredFields = getRequiredFields(konnector)
+  const hasValuesForRequiredFields =
+    intersection(Object.keys(values), requiredFields).length ===
+    requiredFields.length
+
+  return hasValuesForRequiredFields
+}
+
 export default {
   defaultFieldsValues,
   getIdentifier,
   sanitize,
   sanitizeFields,
-  getFieldsValues
+  getFieldsValues,
+  getRequiredFields,
+  hasValuesForRequiredFields
 }
